Network.TypedProtocol.Codec.CBOR

Documentation

mkCodecCborLazyBS

Arguments

:: forall ps m. MonadST m

=> (forall (st :: ps) (st' :: ps). StateTokenI st => ActiveState st => Message ps st st' -> Encoding)
cbor encoder

-> (forall (st :: ps) s. ActiveState st => StateToken st -> Decoder s (SomeMessage st))
cbor decoder

-> Codec ps DeserialiseFailure m ByteString

Construct a Codec for a CBOR based serialisation format, using lazy ByteStrings.

This is an adaptor between the cborg library and the Codec abstraction.

It takes encode and decode functions for the protocol messages that use the CBOR library encoder and decoder.

mkCodecCborStrictBS

Arguments

:: forall ps m. MonadST m

=> (forall (st :: ps) (st' :: ps). StateTokenI st => ActiveState st => Message ps st st' -> Encoding)
cbor encoder

-> (forall (st :: ps) s. ActiveState st => StateToken st -> Decoder s (SomeMessage st))
cbor decoder

-> Codec ps DeserialiseFailure m ByteString

Construct a Codec for a CBOR based serialisation format, using strict ByteStrings.

This is an adaptor between the cborg library and the Codec abstraction.

It takes encode and decode functions for the protocol messages that use the CBOR library encoder and decoder.

Note that this is less efficient than the mkCodecCborLazyBS variant because it has to copy and concatenate the result of the encoder (which natively produces chunks).

convertCborDecoderBS

Arguments

:: forall s m a. Functor m

=> Decoder s a
cbor decoder

-> (forall b. ST s b -> m b)
lift ST computation (e.g. stToIO, stToPrim, etc)

-> m (DecodeStep ByteString DeserialiseFailure m a)

convertCborDecoderLBS

Arguments

:: forall s m a. Monad m

=> Decoder s a
cbor decoder

-> (forall b. ST s b -> m b)
lift ST computation (e.g. stToIO, stToPrim, etc)

-> m (DecodeStep ByteString DeserialiseFailure m a)

Re-exports

data DeserialiseFailure

Constructors

DeserialiseFailure ByteOffset String